Much Much Better
But not functional yet. Just the basics. I still have to create
the shuffle function and I still have to find a way to track which
cards are flipped.

Also I should really decouple and encapsulate with modules in
different files. This is a small enough app that I wouldn't really
need to do that but I want to do it for practice.